+++ index.php
@@ -4,11 +4,6 @@ require_once 'bootstrap.php';
 
 $showCompleteTasks = rand(0, 1);
 $connection = connection($config['dbWork']);
-session_start();
-
-if (getParam($_GET, 'user') == 'exit') {
-    exitSession();
-}
 
 if ($_SESSION) {
