+++ logout.php
@@ -1,0 +1,4 @@
+<?php
+require_once 'bootstrap.php';
+
+logout();
